+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.4.1
 
+* Versions:
+  * `Coroutines`: `1.4.1` -> `1.4.2`
+  * `Klock`: `2.0.0` -> `2.0.1`
 * Add `CollectionKronScheduler`. It will give opportunity to unite several schedulers in one
 
 ## 0.4.0